git-svn-id: svn://svn.debian.org/svn/pkg-samba/branches/samba/experimental@3551 fc4039ab-9d04-0410-8cac-899223bdd6b0

+++ source3/lib/tdb_validate.c
@@ -192,7 +192,7 @@ int tdb_validate_open(const char *tdb_path, tdb_validate_data_func validate_fn)
 
 	DEBUG(5, ("tdb_validate_open called for tdb '%s'\n", tdb_path));
 
-	tdb = tdb_open_log(tdb_path, 0, TDB_DEFAULT, O_RDONLY, 0);
+	tdb = tdb_open_log(tdb_path, 0, TDB_DEFAULT, O_RDWR, 0);
 	if (!tdb) {
 		DEBUG(1, ("Error opening tdb %s\n", tdb_path));
 		return ret;
@@ -281,6 +281,11 @@ static int tdb_backup(TALLOC_CTX *ctx, const char *src_path,
 	}
 
 	tmp_path = talloc_asprintf(ctx, "%s%s", dst_path, ".tmp");
+	if (!tmp_path) {
+		DEBUG(3, ("talloc fail\n"));
+		goto done;
+	}
+
 	unlink(tmp_path);
 	dst_tdb = tdb_open_log(tmp_path,
 			       hash_size ? hash_size : tdb_hash_size(src_tdb),
@@ -355,6 +360,10 @@ static int rename_file_with_suffix(TALLOC_CTX *ctx, const char *path,
 	char *dst_path;
 
 	dst_path = talloc_asprintf(ctx, "%s%s", path, suffix);
+	if (dst_path == NULL) {
+		DEBUG(3, ("error out of memory\n"));
+		return ret;
+	}
 
 	ret = (rename(path, dst_path) != 0);
 
@@ -394,6 +403,10 @@ static int tdb_backup_with_rotate(TALLOC_CTX *ctx, const char *src_path,
 	{
 		char *rotate_path = talloc_asprintf(ctx, "%s%s", dst_path,
 						    rotate_suffix);
+		if (rotate_path == NULL) {
+			DEBUG(10, ("talloc fail\n"));
+			return -1;
+		}
 		DEBUG(10, ("backup of %s failed due to lack of space\n",
 			   src_path));
 		DEBUGADD(10, ("trying to free some space by removing rotated "
@@ -450,6 +463,10 @@ int tdb_validate_and_backup(const char *tdb_path,
 	}
 
 	tdb_path_backup = talloc_asprintf(ctx, "%s%s", tdb_path, backup_suffix);
+	if (!tdb_path_backup) {
+		DEBUG(0, ("tdb_validate_and_backup: out of memory\n"));
+		goto done;
+	}
 
 	ret = tdb_validate_open(tdb_path, validate_fn);
